+++ cli/tsc/dts/lib.deno.window.d.ts
@@ -5,6 +5,7 @@
 /// <reference lib="deno.shared_globals" />
 /// <reference lib="deno.webgpu" />
 /// <reference lib="deno.webstorage" />
+/// <reference lib="deno.canvas" />
 /// <reference lib="esnext" />
 /// <reference lib="deno.cache" />
