+++ ext/web/lib.deno_web.d.ts
@@ -1237,31 +1237,3 @@ declare var DecompressionStream: {
 declare function reportError(
   error: any,
 ): void;
-
-/** @category Web APIs */
-type PredefinedColorSpace = "srgb" | "display-p3";
-
-/** @category Web APIs */
-interface ImageDataSettings {
-  readonly colorSpace?: PredefinedColorSpace;
-}
-
-/** @category Web APIs */
-interface ImageData {
-  readonly colorSpace: PredefinedColorSpace;
-  readonly data: Uint8ClampedArray;
-  readonly height: number;
-  readonly width: number;
-}
-
-/** @category Web APIs */
-declare var ImageData: {
-  prototype: ImageData;
-  new (sw: number, sh: number, settings?: ImageDataSettings): ImageData;
-  new (
-    data: Uint8ClampedArray,
-    sw: number,
-    sh?: number,
-    settings?: ImageDataSettings,
-  ): ImageData;
-};
